Separate out the LOCKING and RETRYING memory wrappers in the initialization code so that the 'locking' wrapper can be used alone when multi-threaded rendering is used.
DETAILS:

The gsmalloc.c and gsmemlok.c are changed to separate out the 'retrying'
and the 'locking' wrappers, since multiple threading doesn't need the
'retrying'. NB, 'retrying' is only used by the clist logic to to "cycling"
or "reversion" where the incomplete page is rendered to a (compressed)
raster area, then used as the background for subsequent clist rendering.
This is problematic for color devices when non-idempotent RasterOps are used,
and all but infeasible for PDL's that use PDF or XPS transparency, so the
'retrying' support can probably be dropped.

EXPECTED DIFFERENCES:

None. (locking and retrying wrappers are not ordinarily used).



git-svn-id: http://svn.ghostscript.com/ghostscript/branches/mtrender@8533 a1074d23-0009-0410-80fe-cf8c14f379e6