Fixes for COMPILE_INITS=1. Tested on Windows and linux with AFS and UFST font scalers. Note that this requires/expects gs rev 8558.
As discussed with Henry, this does NOT currently fetch PCL fonts from the
%rom%ttfonts/ path even though they are built into the rom filesystem.
The code changes needed for this to work are too risky for this close to
release.

Also the language_switch build does not set the FontResourceDir as expected
to %rom%Resource/Font/ even though the GenericResourceDir is set correctly
to %rom%Resource/ which will be addressed separately.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@3029 06663e23-700e-0410-b217-a244a6096597