Fix (clist) : Crop transparencsy commands while clist writing, step 1.
DETAILS :

This is a preparation to the next step.
In general, the behavior doesn't change,
but implementation details differ.

The old code maitains a flag cropping_by_y in the clist writer,
which triggers a cropping of objects by the Y axis.
The old code isn't perfect, because
it first crops by [cdev->cropping_min, cdev->cropping_max],
and then by [0, cdev->height].

In the new code we unite both croppings into a single one.
For doing that we remove the flag, and provide
the right interval [cdev->cropping_min, cdev->cropping_max]
in any case. Initially it is [0, cdev->height},
and it is narrowed when a shading bbox is set.
To restore the initial cropping after the shading
is complete, we maintain 2 new fields
save_croping_min and save_cropping_max.

So now the cropping mechanizm does not depend
on shading, and clients may set a croping when they need.
We'll use this feature in the next patch.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8581 a1074d23-0009-0410-80fe-cf8c14f379e6