Fix (graphics) : Stroke width was wrong for short segments.
DETAILS :

Bug 689742 "Nearly missing dotted lines with anisotropic resolution".

Rather the bug title mentions anisotropic rendering,
the problem appears with isotropic one as well.
The bottom of the problem is that the function width_is_thin
used the dash longitude to know its direction,
which gives an incorrect (zero) result
for vertical segments of zero length.

That method became obsolete since revision 7706,
which intorduced the precise dash direction is a special data field.
Using it now for the right computation of the width.

EXPECTED DIFFERENCES :

012-05.ps at 72 dpi.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8614 a1074d23-0009-0410-80fe-cf8c14f379e6