Fix memory leak observed with DEBUG=1 build on mtrender branch. The 'levels' and 'bit_data' of halftones were being leaked when halftones were set by the clist reader.
DETAILS:

Since the 'dht.rc' was filled with 0's, the rc.memory would cause the
gx_imager_dev_ht_install to copy the order (mem_diff would be true).
The gx_ht_copy_ht_order would make a copy, but the one passed in
was never freed. Since the 'src' component information was cleared
during the install, gx_ht_read_and_install needs to save a copy of
the components so that the orders can be released.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8714 a1074d23-0009-0410-80fe-cf8c14f379e6