Allow some compile time constants that affect performance to be set as /D compiler flags (via XCFLAGS makefile macro).
MAX_BITMAP_PATTERN_SIZE controls when the clist pattern accumulator is used,
and on machines with a decent amount of RAM, the default of 1 Mb is too
small and causes the slower clist method to be used too often. Setting this
to 32 Mb is probably better on hosts.

The GS_CLIENT_COLOR_MAX_COMPONENTS default of 252 is needed for Adobe CPSI
compatibility and for RIPS that want to use a large number of DeviceN
colorants, but this increases the size of some allocations and slows down
performance and is much too large for printers. Setting this to 6 or 8 is
probably plenty.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8719 a1074d23-0009-0410-80fe-cf8c14f379e6