Fix (stroking) : Prevent unpainted gaps between neighbour strokes that could appear due to stroke adjustment.
DETAILS :

Bug 687974 "pdf file has banding with ghostscript but not with acrobat".

The test case includes a gradient, which is represented
as a set of paralel strokes. Due to stroke widths equals to fractional pixels,
stroke adjustment shifts some of them so that
an unpainted gap appears between strokes and the gradient looks striped.
Note the test document is created with Quark Express 4.11 .

This patch implements a hewristic recognizer for parallel contacting strokes
and suppresses stroke adjustment for them. See comments in code
for more details.

Note that the recognizer data are associated with a device but they
are not a device property. We store them in a device structure because
we haven't got an associated structure in the graphic state,
and we don't want to introduce one now against extra complexity
with memory management. Also the recognizer code
Appears to be distributed along multiple functions and we would like
to apply a better incapsulation. Nevertheless the encapsulation
would cause CPU time expense for multiple small function calls
unless we create new macros. So commonly we're unhappy of
this code but we don't see a better solution.

Also note that the recognizer data is copied from/to clipper
device when the stroking algorithm installs and deinstalls a clipper.
So the data actually is not a property of a device.
Also they're not a part of graphic state because they must live
across gsave/grestore (the test case requires so).

EXPECTED DIFFERENCES :

Minor raster difference (a 1 pixel shift of some strokes) at 72 dpi :

"Altona-Testsuite_p2_S_x3.pdf"
"Altona_Visual_bb_1v1_x3.pdf"
"Altona_Visual_sb_1v1_x3.pdf"

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8790 a1074d23-0009-0410-80fe-cf8c14f379e6