Fix (clist writer) : Indeterminizm writing a degenerate curve.
DETAILS :

Bug 689911 Shifting raster data, Windows MSVC 8 non-debug build.

In function cmd_put_path the variable 'prev' was wrongly localized
in the loop body when need to save/pass data from the previous cycle.
It worked fine with debug build, because the compiler
always commits variables to RAM for debug purpose.
With release build the variable 'prev' never changes
because formally speaking the program never uses its value
after it is initialized.

The bug persists since the repository was created on March 9 2000.
The old behavior may be indeterministic.

With MSVC8 the effect is occasionaly visible when a degenerate curve
{x y x y x y x y} is written into clist.
See the simplified test case attached to the bug report.

EXPECTED DIFFERENCES :

None, but the old behavior may be indeterminiztic.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8815 a1074d23-0009-0410-80fe-cf8c14f379e6