Fix a hang in s_jpxd_process. Bug 688807.
DETAILS:

The problem file contains a CMYK JPX stream, which jasper does not
recognize, treating it as sRGB with an extra component. This causes
us to try an return 3-component data, and end up looping forever if
only a fraction of the expected pixel is available at the end of a
line.

Check for this event and return an error instead.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8817 a1074d23-0009-0410-80fe-cf8c14f379e6