Fix (pdfwrite): Images ina DeviceN space with more than 60 inks caused a crash.
Details:
Bug #689864 "Regression: 09-34.PS core dumps with pdfwrite"

When emitting an image, using the Flate encoder, the number of colorants in
the DeviceN space was not checked against the maximum permitted. Since the
encoder uses a fixed buffer whose size depends on the maximum number of inks, this
could cause a write beyond the end of the buffer.

(spngpx.h) Increase the maximum number of colorants to 256
(spngp.c) Return error if the number of colorants in the space
          exceeds the maximum permitted.

I can't check the generated PDF file, as no PDF reader seems able to open a PDF
file with more than 32 inks, the architectural limit in the current version
of Acrobat.

EXPECTED DIFFERENCES:
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8823 a1074d23-0009-0410-80fe-cf8c14f379e6