Fix (graphics) : Write tile id into a pattern clist.
DETAILS :

Bug 689818 "Regression: Test file imagemask-with-pattern.ps fails with banding and pbmraw ".

The old code doesn't write a tile id when writing the tile to clist,
because the clist interpreter doesn't use tile id.
This causes a failure after we implemented clist-based patterns.
While playing back a pattrn clist to a page clist writer device,
the clist writer does need a tile id
for choosing an entry of an internal hash table.

Actually the failure happens due to a strange check
"tile->id != gx_no_bitmap_id" in
clist_strip_tile_rectangle in gxclrect.c ln 479.
It is not clear what it was intended for.
When a tile id is zerfo, the check skips writing tile size to clist,
so that a zerodevide happens during the clist playback.

We decided to provide meanful ids
because the clist writer needs them anyway.
Fixing the check to be a subject of a separate patch
(if a fix is really necessary).

This patch simply writes tile id when the clist
is for a pattern. A new nacro IS_CLIST_FOR_PATTERN(cdev)
Defines the condition for that.
Some internal functions change prototypes
with adding a new boolean argument for_pattern to know
whether a clist is for a pattern.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8862 a1074d23-0009-0410-80fe-cf8c14f379e6