Fix for Bug 689246.  When performing interpolation, the interpolation is now performed in the source colorspace as opposed to the device space.  Custom color callback is no longer bypassed by direct concretization as it was previously.
DETAILS:

Previous code would take device color 8 bit values and do the interpolation.  That portion of the code is the same.   For cases where we have nondevice color the handling is now different.  Indexed colors are mapped to their base space and interpolated.  This is followed by a remap operation.  Device independent colors are interpolated followed by a remap operation.

Note that for cases where there is a significant enlargement occurring and we are not in a device color space, there can be a performance cost due to current inefficiencies in our color remapping operations.  This can be costly if you have a CMYK CRD MLUT for example.  There is currently a project underway to improve the efficiency of the color flow.  For the current time, if this is an issue, improved speed can be achieved by invoking -dNOINTERPOLATION

EXPECTED DIFFERENCES:

148-11.PS
12-07B.PS
23-32.PS
13-03.PS
13-12.PS

The first two files have interpolation defined as on for some of their images. Hence there will be differences.  With this change we more closely match the AR output for 12-07B.PS.  148-11.ps  has only a minor difference in its coloring due to the interpolation in the different color space.  The other files are nondeterministic.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8868 a1074d23-0009-0410-80fe-cf8c14f379e6