Fix (True Type font loader) : Can't work around unsorted 'loca' when 'loca' is not loaded.
DETAILS :

Bug 689893 "Regression: CID font emulation is broken."

A work around against unsorted 'loca' and its recent improvements
assume that 'loca' presents in sfnts. It is not true
when a CID font is emulated with a True Type disk font,
because in that case the font machinery delay loading
'loca' and 'glyp' elements until a real request from a document.
For more details see comments added in code.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8881 a1074d23-0009-0410-80fe-cf8c14f379e6