Fix (clist) : Skip writing unuseful data with linear color triangle.
DETAILS :

This fixes a minor problem what was discovered while working on
bug 689918 "Ghostscript 8.62 and 8.63 segmentation fault".

cmd_write_trapezoid_cmd wrote an unused data to clist when
processing a linear color triangle. The variables ybot, ytop are never used
while a clist interpretation. In same time the data are not initialized
in that case, so they imply an indeterministic content and length of the clist.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8882 a1074d23-0009-0410-80fe-cf8c14f379e6