Fix __func__ on the default (non-C99) gcc build.
Details:

__func__ is a variable in C99. The earlier __FUNCTION__ is a magic
preproccessor define in MSVC 7 and later, so the #if defined() clause
enabled it. However, in GCC it is a magic string literal instead and so
this clause did not enable it. Therefore we add an explicit check for
the appropriate GCC version and define the same substitution for that
compiler.

Differences:

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9001 a1074d23-0009-0410-80fe-cf8c14f379e6