Replace an abs() call with our any_abs() macro. Also remove an unnecessary math_.h include.
Details:

A call to abs() was added in r8942. Technically the prototype for this
is in stdlib.h, but it's a built-in in most compilers so this didn't
raise a warning. We do include stdlib.h elsewhere, usually for qsort()
or malloc(), but abs() is only called from comment pseudo code; in
practice the any_abs() macro or fabs() from math.h are used instead.
Because this isn't in an inner loop, speed isn't important, so we prefer
the macro to a system header include for consistency.

Sine the beginning of repository history, gdevbit.c has included math.h,
but I don't see any math library calls now, so it as been removed.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9007 a1074d23-0009-0410-80fe-cf8c14f379e6