Silence an 'uninitialised variable' warning from Coverity.
Details:
Function initialisation is a bit round about. We pass a function pointer to
make_sampled_function, which assigns it to a local variable. That variable is passed to
gs_function_Sd_init, which immediately zeros it....

Since the pointer is never actually used, setting it to NULL should silence the warning.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9008 a1074d23-0009-0410-80fe-cf8c14f379e6