Document the motivation for a source style item.
Details:

There has been some controvesy over the rule that single-statement
bodies of control structs must be on their own line. Specifically:

  if (code < 0)
    return code;

instad of:

  if (code < 0) return code;

despite the latter being less obtrusive for common error checks.

Several developers have requested this because it allows setting a
breakpoint directly on the return statement instead of a conditional
return on the conditional line, or a breakpoint at the return inside
the single source line. This is therefore a practical issue, not just
one of readability and style.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9034 a1074d23-0009-0410-80fe-cf8c14f379e6