Fix SEGV when NumRenderingThreads > 0 caused by improper clist reader phase initialization. Bug #690108.
DETAILS:
Make sure that the clsit device is initialized to 'reader' mode before checking
the 'render_threads' element of the reader device. Prevents a SEGV when the
corresponding element of the 'writer' device structure was non-zero at the end
of the page since the render_threads table will not have been allocated and
some other (invalid) pointer would be used instead.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9128 a1074d23-0009-0410-80fe-cf8c14f379e6