Setting of the "default" CMYK, RGB, and Gray to Device CMYK, RGB, and Gray when an Smask occurs even if UseCIEColor has been specified.
DETAILS:

This is how AR behaves. It ends up using device CMYK for example to compute luminance of the soft mask by L = (0.30 (1-C) + 0.59 (1-M) + 0.11 (1-Y)) * (1-K)  even though it uses SWOP CMYK for the rendering of CMYK in the color fills.  The approach used here is that the if we are in a soft mask we will define the default CMYK to be device CMYK even if UseCIEColor as been set to true.  Upon completion of the softmask, the default CMYK will be return to its previous setting.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9146 a1074d23-0009-0410-80fe-cf8c14f379e6