Addition of function to compute the soft mask luminosity value.  The computation is different depending upon if the source buffer data is device gray, device RGB, device CMYK or CIE based.  Also included is some minor debug code for dumping data buffers.
DETAILS:

As detailed in the PDF specification:
For device color spaces, convert the color (in the Soft mask buffer) to DeviceGray by device-dependent means and use the resulting gray value as the luminosity, with no compensation for gamma or other color calibration. This method makes no pretense of colorimetric correctness; it merely provides a numerically simple means to produce continuous-tone mask values. Here are some recommended formulas for converting from DeviceRGB and DeviceCMYK, respectively

L = 0.30 R + 0.59 G + 0.11 B
L = (0.30 (1-C) + 0.59 (1-M) + 0.11 (1-Y)) * (1-K)

Following this conversion, the result is passed through a separately specified transfer function, allowing the masking effect to be customized.

Note that according to the specification:
For CIE-based spaces, convert to the CIE 1931 XYZ space and use the Y component
as the luminosity. This produces a colorimetrically correct luminosity.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9148 a1074d23-0009-0410-80fe-cf8c14f379e6