Addition of color mapping procs in pdf14 device which will be used when a soft mask is pushed.
DETAILS:
This is necessary since the standard color mapping procs rely upon the device process color model.  When we are performing operations within a softmask group, we should NOT be mapping to the device color model but rather to the color space that has been defined for the soft mask transparency group.

EXPECTED DIFFERENCES:
None.  This commit only introduces the new procs.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9151 a1074d23-0009-0410-80fe-cf8c14f379e6