Addition of structures to assist in switching the color mapping procs during the soft mask rendering.
DETAILS:

When the soft mask is pushed we need a structure to maintain the information that we will be changing due to the fact that we should be rendering in the color space defined by the soft mask transparency group.  When the soft mask is popped we will restore this information.

EXPECTED DIFFERENCES:
None.  This only defines the structures.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9152 a1074d23-0009-0410-80fe-cf8c14f379e6