Fix for bug 689961.
There was a problem with the testing to determine if a color space mapping was linear within a triangle and along a line.  The old code assumed the number of components was the same in the client and device color space.  In general this is not true, especially if we have a DeviceN color space with one color and a CMYK output device, which was the test case given.

EXPECTED DIFFERENCES:

This depends very much on the output device color space, the source color space and the resolution.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9159 a1074d23-0009-0410-80fe-cf8c14f379e6