Fix (clist writer) : Wrong compositor cropping when a shading is entirely clipped out (continued).
DETAILS :

Bug 690106 "Regression: /unregistered in --run--"

The recent patch for the subject is not sufficient because
an empty clipping condition cannot be anticipated in the clist writer.
To fix it we introduce a new flag gx_device_clist_writer_s::cropping_saved
especially for clist_fill_path to know whether it was called recursively
for passing an untrivial clipping path for a shading. The cropping must be restored
in the outer call iff it is set.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9168 a1074d23-0009-0410-80fe-cf8c14f379e6