Changed decleration of det from float to double in gs_matrix_invert().
DETAILS:

On peeves the results of gs_matrix_invert() change depending on
whether or not compiler optimization is enabled if det is declared
as a float.  I believe this is due to the optimized code never
storing the det value so it never get shortened to a float; by
declaring it as a double the unoptimized code should act the same.

EXPECTED DIFFERENCES:

Since the nightly regressions are run with optimization enabled
there shouldn't be any differences.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9181 a1074d23-0009-0410-80fe-cf8c14f379e6