Fix for bug 690103.  This issue was introduced with r8868 where spatial interpolation in the source color space was introduced.  The part that was missing in that commit was to handle the case where we had 8 bit image data that was non CIE based and NOT encoded with [0 1] .
DETAILS:

This patch was tested with forced interpolation on a set of files that are known to have issues with interpolation. It performed properly on those and on the customer's bug file.

The part that was missing in the old code was the case where we have encoded
(not [0 1]) nonindexed 8 bit image data that is not CIE based.	For this case,
it was not getting properly decoded prior to interpolation.

EXPECTED DIFFERENCES:

None for the standard regression.  With forced interpolation it will behave correctly now for the case of 8 bit image data that is non indexed, non CIE and encoded in something other than the standard range of [0 1].

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9184 a1074d23-0009-0410-80fe-cf8c14f379e6