Added support for doing blending in the color space specified for the transparency group.
DETAILS:

The process for doing this is same as the process that was used for adjusting for the softmask color space.  Note that new member variables were added to gs_transparency_group_params_s and pdf14_buf_s .  These will need to be handled on the clist writer and reader side to ensure that the needed information is passed along between gs_begin_transparency_group (prior to the clist writer) and gx_begin_transparency_group (during the reading of the clist que).

The buffers are now allocated based upon the color space size defined by the transparency group.  Also, if there is a mismatch between the number of color components during the popping of a group, the buffer is transformed to the proper color space.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9209 a1074d23-0009-0410-80fe-cf8c14f379e6