Replace the awki cgi interface with static file creation through a GNU Makefile. All pages are now generated directly by parser.awk, which is called directly by the makefile.
In short, type 'make' to generate the static website from the .awki
source files.

The cgi functionality is preserved in a reduced search.cgi which only
implements the search function provided by the original wiki.

The makefile also generates a 'contents' page with the assistance of
listawki.awk which converts of list of .awki filenames into an
awki-syntax page which parser.awk can render.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9266 a1074d23-0009-0410-80fe-cf8c14f379e6