Fix for weird math issue under AIX that cause problems in glpyh spacing when writing type42 fonts for ps2write and pdfwrite.
This fixes bug 690118.

Details:

For reasons that I don't undertand under AIX (ushort) (double * double)
produces a 0 if the first double is > -1 and < 0 (and the sum is < -1).
Presumably the cast happens before the multiplication, but this isn't
reasonable and doesn't occur on other operating systems, even running
the same version of gcc.  Temporarily assigning the sum to a float and
then casting to a ushort resolves the problem.

The temporary assignment should be removed by the optimizer, so there
shouldn't be a performance hit (I did check, the optimized code is still
correct under AIX).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9277 a1074d23-0009-0410-80fe-cf8c14f379e6