Fix (PDF interpreter) : Avoid extra erasepage, part 1.
DETAILS :

This is the first partial fix for the
bug 690131 "Optimize consecutive setpagedevice".

The PDF interpreter calls extra setpagedevice for initializing the number of spot colors.
First it happens in pdfshowpage_setpage, and we merge that call
with another one that sets other multiple page parameters.
Second in pdfshowpage_finish we replace it with .setpagedevice,
which doesn't call .postinstall that calls erasepage.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9278 a1074d23-0009-0410-80fe-cf8c14f379e6