Fix (pdfwrite): Fonts with unusual FontMatrix generating incorrect FontBBox.
Details:
bug #689879 "wrong FontBBox in PDF file"

Fonts other than type 3 may not have a FontMatrix in PDF, these fonts are assumed to be on a 1000x1000 design grid. Most fonts already are, and embedded TrueType fonts are already catered for, but some unusual fonts have a non-standard (not [0.001 0 0 0.001 0 0]) FontMatrix. The matrix is accounted for when scaling the font, so the Tf operator is correct in the page content stream, but the FontBBox was calculated using the unscaled glyph outline.

We now scale the font outline into a 1000x1000 design grid before calculating the bounding box, so that the FontBBox entry will be correct.

Expected Differences
None (FontBBox only used to highlight text selection in Acrobat)

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9284 a1074d23-0009-0410-80fe-cf8c14f379e6