Fix (graphics) : Redundant patterns in clist (continued 4).
DETAILS :

This is a further improvement for the bug 690133
"Regression: Ghostscript takes up all disk space when rendering eog or GIMP PDF output in 1200 dpi".
Also it is a preparation for a partial fix for the bug 690131 "Optimize consecutive setpagedevice".

This patch passes the coverage rectangle of a graphic element
for a better computation in cmd_put_drawing_color of the size
would be consumed while serializing a color to clist
with copying to all covered bands.
When we need to force writing to all bands,
we pass a NULL instead the coverage rectangle.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9285 a1074d23-0009-0410-80fe-cf8c14f379e6