Fix (graphics) : Redundant patterns in clist (continued 5).
DETAILS :

This is a further improvement for the bug 690133
"Regression: Ghostscript takes up all disk space when rendering eog or GIMP PDF output in 1200 dpi".
Also it is a preparation for a partial fix for the bug 690131 "Optimize consecutive setpagedevice".

The last partch wrongly distributed the saved pattern id for all bands
when writing a non-pattern color to all bands.
We believe it never happens the current revision with reasonable banding settings,
but fix fix it for code consistency and for avoiding problems
with further revisions.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9286 a1074d23-0009-0410-80fe-cf8c14f379e6