Fix (graphics) : Redundant patterns in clist (continued 6).
DETAILS :

This is a further improvement for the bug 690133
"Regression: Ghostscript takes up all disk space when rendering eog or GIMP PDF output in 1200 dpi".
Also it is a preparation for a partial fix for the bug 690131 "Optimize consecutive setpagedevice".

The last partch appears syntactically incorrect, now fixing.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9287 a1074d23-0009-0410-80fe-cf8c14f379e6