Add the CVS directory to the romfs exclusion list. Bug 690205.
Details:

Currently the mkromfs script just includes the entire file tree under
Resource/ but the presence of any extraneous directories there confuses
the enumeration machinery. We have a exclusion for the .svn directory
so builds out of svn work, but apparently some people have a CVS mirror?

Anyway, it isn't difficult to add an additional exclusion for CVS
metadata. This isn't an issue for git, darcs or mercurial because they
keep all their repository data in the top-level directory.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9293 a1074d23-0009-0410-80fe-cf8c14f379e6