Fix (pdfwrite): Image with SMask fails when inside transparency group.
Details:
Bug #690139: "Regression: pdfwrite creates empty file".

The PDF Interpreter internally surrounds SMask images with an extra transparency group definition. If emitted to PDF this is invalid and causes Acrobat to complain. This was fixed in revision 6878 (bug 688767) by adding an extra parameter  'Image_with_SMask' to the dictionary.

The pdfwrite code checks for the existence of this in the group, and if present (and true) skips the transparency group begin. It also sets a flag in the device, to skip the end of the group as well.

However, the flag is only reset when a new transparency group is begun. If the SMask was nested inside a group then the flag remained set, causing the 'end' for all enclosing groups to be skipped as well. This caused empty output or, in extreme cases, a 'rangecheck' error in discardtransparencygroup, because we overflowed an internal stack.

We now set the device flag to false after we have skipped an end group.

EXPECTED DIFFERENCES:
Bug688601a.pdf
Bug689918.pdf
test file for bug 689805

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9303 a1074d23-0009-0410-80fe-cf8c14f379e6