Merge of trunk changes 9175-9326 into smask branch.
DETAILS:

I am getting ready to commit a patch related to transparency and overprint into the trunk and I want to bring that in the soft mask too, as some files have both overprint and transparency.  I makes sense to just do a full merge prior to that to enable easier debug and conflict resolution.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9327 a1074d23-0009-0410-80fe-cf8c14f379e6