Added target devices color_info.num_components to the color procs stack to enable proper implementation of shading code during clist writing.
DETAILS:  The shading code during clist writing wants to handle everything in the target device color space.  This should not be the case if for example you are drawing a transparency group or soft mask in a particular blending space.  All the shading code uses the target device's color_info.num_components to determine the color space.  By updating the target device color_info.num_components like we do with the pdf14 device we obtain the proper rendering.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9379 a1074d23-0009-0410-80fe-cf8c14f379e6