Remove $(GS_LIB)\Resource directory from the search path in the msvc build to shadowing the compiled in version.
GS_LIB is set in the Windows registery by the installer, but it does not
currently install a duplicate copy of the Resource tree in the
filesystem. Having this directory in the search path caused Ghostscript
to use it for Resource lookup to the exclusion of the %rom% version, and
since the files weren't installed there, Ghostscript would error out in
the initialization phase.

Files like cidfmap and available fonts can still be overridden by
putting them in $(GS_LIB)\lib and $(GS_LIB)\fonts.

Patch from Russell Lang.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9404 a1074d23-0009-0410-80fe-cf8c14f379e6