Fix to ensure shading fills do NOT occur in the device color space when we are within a transparency group with a different color space.
DETAILS:  The previous fix for this was to alter the number of color components in the color info of the target device.  This is not a desirable approach since that value really should be private to the device.  A better approach is to introduce into the imager state a boolean that indicates if we are within a transparency group.  When the pattern color is created for the shading code and we pass through the pdf14clist fill path we now set this value true if we are within a transparency group.  The shading code has a check called is_linear_color_applicable which it uses to determine if it can do the shading in the target device color space.  If the boolean flag related to transparency is true, then it will keep the colors in the source space. To handle it in the device space would require knowing all the nested transparency group color spaces as well as the transparency values and blending operations.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9428 a1074d23-0009-0410-80fe-cf8c14f379e6