Fix (pswrite): Incorrect A3 page size
Details:
bug #690259 "A3 pagesize pointsize definition wrong in gdevpsu.c"

The definition of the A3 paper size was 842 by 1190 in psw_write_page_header(). The definitions elsewhere in GS have a height of 1191 (which is closer to the ISO metric paper size).

Modified the page height to be 1191, also modified the media matching code so that it allows +/- 5 points when matching media.

Thanks to Roland Friedwagner for debugging the problem and supplying a patch, and to Alex Cherepanov for suggesting the media matching.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9445 a1074d23-0009-0410-80fe-cf8c14f379e6