Fix to achieve proper image color handling with the clist when we have transparency group color spaces that are different than the target device.
DETAILS:  The current code forwards the begin_typed_image and begin_image calls with the target device.  Later during the rect fills, which have no access to the imager state, we end up with no knowledge about the transparency group color space, only about the target device color info which can be different than the trasparency group.  This fix provides this information in two ways.  One is to use the default image calls with the pdf14 device (not the target device) when clist_begin_type_image fails.  The other is to add additional information in the clist writer device about the true color space that the clist writer should be using.  Specifically we have:
Addition of color_info.depth to data stored during transparency group.
Addition of clist_color_info to clist writer device.  This is updated as a transparency group is pushed or popped.  It is initialized with the target device color information.  cmd_put_color in gxclutil.c makes use of this information. Change of pdf14_clist_procs encode and decode color procs to the pdf14 device encode and decode procs. pdf14_clist_begin_typed_image and pdf14_clist_begin_image now use gx_default_begin_typed_image and
gx_default_begin_image with the pdf14clist device NOT the target device IF clist_begin_typed_image can not handle the image itself.  Previously it did use the default procedures but WITH the target device. This caused a loss of information about the updated procs that occur with the pdf14 device during transparency color group changes.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9449 a1074d23-0009-0410-80fe-cf8c14f379e6