Fixes to transparency to recognize Patterns with transparency in a PDF and install the pdf14 compositor (push_pdf14trans_device) when the pattern is rendered in to the pattern_accumulator device. Bugs 688278 for customer 700 and user bugs 690189 and 690422.
DETAILS:

Added a 'uses_transparency' boolean to the Type1 pattern template. which defaults
to 0 (false), set true when 'patternusestransparency' is true in the PDF interpreter.

Also, the patternusestransparency is used when opening a PDF to ensure that the
'pageusestransparency' used to install the pdf14 compositor for the page is true.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9458 a1074d23-0009-0410-80fe-cf8c14f379e6