Protect some printf-like calls in the pcl3 output device from spurious format specifiers. This eliminates a security-related gcc warning.
Details:

I believe the previous code was actually safe. However, this approach is
more robust. If any of the string buffers passed to errprintf() happened
to contain a format specifier (like %d) an out of bounds stack reference
could result.

This allows the default build to complete with -Werror=format-security,
which the Mandriva linux distribution is reported to be using.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9460 a1074d23-0009-0410-80fe-cf8c14f379e6