Fix (TrueType interpreter): Ignore a 'faulty font' error
Details:
bug #689471 "Error in Processing Type42 Charstrings"

The font in question is missing a glyph in the GLYF table, which is used
as a component by a composite glyph. A later instruction then tries to access
a non-existent point (probably defined in the missing component), which causes
an error.

FreeType 1 and 2 both ignore this error unless pedantic_hinting is set, other
TrueType interpreters behave similarly.

Modified the code in Ins_MIAP to ignore errors of this type, and continue.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9468 a1074d23-0009-0410-80fe-cf8c14f379e6