Fix (TrueType interpreter): Ignore a 'faulty font' condition
Details:
bug #689760 "Missing characters: Failed to interpret TT instructions"

The font in question is technically invalid. The problem occurs in function 1 of the
FPGM table:

	00069: PUSHB[2]              0    21
	00072: JMPR
	00073: ABS
	00074: DUP
	00075: ROLL
	00076: RCVT
	00077: DUP
	00078: ROLL
	00079: SUB
	00080: ABS
	00081: PUSHB[1]             50
	00083: LT
	00084: IF
	00085: SWAP
	00086: EIF
	00087: POP
	00088: ROUND[Black]
	00089: PUSHB[1]             64
	00091: MAX
	00092: ENDF

The JMPR (jump relative) instruction is intended to resume execution at the ENDF
instruction, but the number of bytes is miscounted, falling of the end of the function.
I suspect some TT interpreters catch this by spotting the fact that the IP is past the
end of the function.

We don't seem to be able to check if we've fallen off a function, because we don't store
function lengths, so this code works around the problem by checking the opcode after we
apply the JMPR but befopre resuming execution. If the opcode is *not* ENDF, but the
preceding opcode was, assume that we have a counting problem, and move the IP back one
instruction.

Also applied the same work around to JROT and JROF instructions.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9474 a1074d23-0009-0410-80fe-cf8c14f379e6