Don't strip dynamic driver modules when linking them.
This is better done at the install or packaging stages if space is a
concern.

Also, propagate LDFLAGS to the link command line. This may or may not be
what you want, but is more likely to help helpful than not.

After a patch by Kevin Pyle. Bug 690288. Originally Gentoo bug 215913.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9476 a1074d23-0009-0410-80fe-cf8c14f379e6