Fix for segfault on the file from Bug 690189 caused by not having an open substream for a transparency group when the transparency group ended.
DETAILS:

The 'is_in_page' macro, not to be confused with the 'in_page' element of
the gx_device_pdf structure, was used to skip opening a substream, but
if the stream had not been opened by the time the end_transparency_group
occurred, the pdf_exit_substream would occur anyway. This led to a segfault
when the Pattern accumulation in this file ended since the substream was
closed and the expected pointers were 0.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9494 a1074d23-0009-0410-80fe-cf8c14f379e6