Fix for postscript crashes in regression tests.
DETAILS:  When the PDF14 clist writer device does begin_typed_image it uses the target device through gx_forward_begin_typed_image.  If it cannot handle the image, it uses gx_default_begin_typed_image.  For the PDF14 device this should occur not with the target device but with the pdf14 clist writer device.  I had a return of -1 if the clist_begin_typed_image could not handle the image.  The error was that clist_begin_typed_image is used many places not just from pdf14_clist_begin_typed_image.  The current fix is to make use of the has_transparency marker in the imager state.  If the flag is set, then we return -1 if the clist writer cannot handle the image type, in which case we end up doing a call to gx_default_begin_typed_image but not with the target device.  If the flag is not set,  we use gx_default_begin_typed_image, called from  within  clist_begin_typed_image.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9502 a1074d23-0009-0410-80fe-cf8c14f379e6