Update the msvc build of the freetype bridge for the new link scheme.
This commmit fixes some errors in the earlier untested update. In addition, the earlier FT_LIB (singular) makefile variable is resurrected for the msvc build only. It is necessary because the freetype project file embeds the release number of build variant in the library filename. If FT_LIB is not set it defaults to the multi-threaded debug build of the 2.3.8 release. Obviously this should be changed when building a binary for release, or against a different version of the freetype library.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9503 a1074d23-0009-0410-80fe-cf8c14f379e6