Fix (colour): Faulty error handling
Details:
bug #690312 "gs stalls, never returns on attached figure"

The job causes the 'base' (Gray, RGB, CMYK) colour code to be executed with a completely
full stack, which causes a stackoverflow because this routine needs to push the colour
space name onto the operand stack.

However, the operand stack wasn't being checked until after the continuation procedure
had been pushed onto the execution stack, so after the error handling, we returned to
the continuation procedure, but without an appropriate operand stack. This caused a
loop to run without bounds.

Fixed by checking the operand stack beforesetting up the execution stack for the
continuation and aborting if it is full. Desk checking revealed another potential case
of the same problem, fixed in the same patch.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9519 a1074d23-0009-0410-80fe-cf8c14f379e6