Fix for issue when target device is monochrome and we are using the pdf14 device also now always updating the color mapping procs for all transparency group and soft mask push and pops.
DETAILS:  The previous code was always pushing a 3 channel + 1 alpha device when the PDF14 device was gray.  Now it pushes a 1 channel + 1 alpha device.  Also, we now update the color mapping procs anytime we have a group push.  Previously we were only doing it when the number of channels had changed.  This ensures that we use the pdf14 group color mapping procs when we are in a group which avoids us from mapping to the target device color.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9532 a1074d23-0009-0410-80fe-cf8c14f379e6