Fix (fonts): Potential dereference of NULL pointer.
Details
Flagged by Coverity, the FAPI code used op_show_find to get the text enumerator, but
did not check the return value, which can be NULL for a failure. Later the enumerator
is dereferenced, potentially causing a crash. This is actually almost certainly
impossible since this is a finish routine and the enumerator is known to have existed
previously, but its good to be safe.

Expected differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9604 a1074d23-0009-0410-80fe-cf8c14f379e6