Restore the proper size of the huffman symbol length histogram.
The previous size as 256 KB, but LENCOUNT's highest index is PREFLEN
(or LENMAX - 1, but LENMAX is the largest value of PREFLEN) and that
comes from a byte member of a Jbig2HuffmanEntry. Therefore 256 members
are all that is necessary, or 1 KB on most machines.

This was changed in commit a0531af8 (jbig2dec svn c430) apparently in
the mistaken belief that it needed to scale with the size of the table
itself. In that revision the previous stack allocation of 256 elements
was replaced with the 1k, which later became 64k, generally overflowing
the stack. While 256 elements is a reasonable stack allocation, it's
safer to continue using the heap.