Fix (pdfwrite): Incorrect handling of type 1 counter control hints.
Details:
bug #690349 "corrupt PDF is generated out of PostScript-file with asian fonts"

When converting type 1 outlines to type 2 outlines, we were not catering for
OtherSubrs 12 or 13 (counter control hints) in the first pass over the font.
These operators take a large number of arguments, and these were being left
on the font operand stack. This quickly resulted in a stack overflow and an
invalidfont error. We now remove the operands from the stack when we encounter
them.

In addition, the code for writing CIDFontType0 fonts which called the conversion
routine was not checking the return value, and carried on trying to embed the font
anyway.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9606 a1074d23-0009-0410-80fe-cf8c14f379e6