Enhancement (pdfwrite) : Allow pdfwrite to produce multiple copies of pages. New flag         'DoNumCopies' added.
Details
Bug #690355 : "pdfwrite ignores the "#copies" setting in PostScript input"

The pdfwrite device follows the behaviour of Acrobat Distiller and ignores the /#copies
and /NumCopies settings in PostScript. This is at least partly because these can't be
expected to work properly with Destination annotations specifying a page as the
destination (eg Link or Outline annotations).

However CUPS (and potentially other PDF workflow applications) may have no way of
determining that application-produced PostScript requires multiple copies to be
printed, after the file has been converted to PDF.

For the benefit of such applications a new flag 'DoNumCopies' has been added, if
present then pdfwrite will duplicate each page in the output as many times as the
/#copies or /NumCopies current setting.

There is currently no provision for reordering the pages, the duplicates follow the
original immediately in page order.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9615 a1074d23-0009-0410-80fe-cf8c14f379e6