Fix (pdfwrite): Incorrect handling of short images with subsampling.
Details:
bug #690350 "Conversion to pdf of a color pcl file generated by pjxl300
gsview 4.9 driver"

The PCL file consists of a single large image covering the page. However,
the PCL interpreter calls 'process_zero_rows' (rtraster.c) which has the
side effect (sometimes at least) of closing the image and restarting it.

Pdfwrite already has code to handle a short image like this, it replaces
the image dictionary /Height value with the number of rows of data
actually received. However, if subsampling is enabled, this can be
incorrect, some of the rows of received data may have been discarded.

Altered pdf_end_image_binary to check if subsampling is taking place and
if it is to calculate the subsample factor and divide the number of rows
of data received by the subsample factor before storing as the new /Height.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9618 a1074d23-0009-0410-80fe-cf8c14f379e6