Fix for crash of bug688728.pdf when going to pdfwrite.
DETAILS:

Looking at bug689422.pdf with pdfwrite it was clear that when we had a pattern that had transparency pdfwrite was writing out the inline image AND the resource for the image which would get called with a Do (this made the inline image show through in transparent regions when the resource image with the softmask was drawn).  Having pdf_begin_typed_image_impl generate a resource and not an inline image fixes the issue.  bug689422.pdf still has a geometric issue when going to pdfwrite, but this is also present in the trunk and is not related to transparency.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9625 a1074d23-0009-0410-80fe-cf8c14f379e6