Fix for crash on 18-02B in clist code. The clist cbuf (on the stack) would write past the end of the buffer are polluting the next varaible on the stack (data_bits). Fixes 690399.
DETAILS:

This problem was introduced with r8712 (leonardo). The return value for
gx_dc_pattern_read_raster should be the number of bytes consumed, but in some
(rare) cases it would return 0, causing a later contamination of a byte past the
end of a buffer on the stack, polluting another value. The effect would be
dependent on the 'endian-ness' since only one byte was overwritten.

Also 'protection' for this condition was added in top_up_cbuf to prevent
future problems.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@9628 a1074d23-0009-0410-80fe-cf8c14f379e6