Prevent the 'extra image' problem with pdfwrite. This fixes 689422.pdf with the pdfwrite device (not the scaling issue -- just the black vs. white in the pattern). This area needs to get cleaned up since with this change I suspect the 'skip image' hack is now not needed.
DETAILS:

This is a sort of rushed fix to let us get the smask branch into the trunk,
but I think a better approach that testing for 'currentdevice pdfwrite eq'
would be to allow pdfwrite (or other HL devices that handle images with
SMask to return an (unique) error code to skip the image. Another possible
approach (that I like somewhat less) is to have devices set a boolean
parameter if they can handle SMask so we can ask the device.

IMO the other hackish modification in 'doimage' should also get fixed to
not check for 'pdfwrite'.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/smask_work@9636 a1074d23-0009-0410-80fe-cf8c14f379e6