Check the return value of sscanf when parsing pbm dimensions.
This is a follow-on to the previous commit. It shouldn't be
possible to reach the sscanf with less than one digit in the
parse buffer, but we abort anyway in case sscanf can't find
anything. If the code could proceed from that point with an
uninitialized value in the dim array the call to jbig2_image_new()
could request excessive resources.